refactor(#3479): extract idle-tail offset-resolution cluster from tui_prompt_relay
Behavior-preserving move of 4 `#[cfg(unix)]` helpers (resolve_idle_tail_start_offset,
claude_idle_response_start_offset_after_timestamp, normalize_transcript_fallback_offset,
clamp_idle_tail_start_offset_to_committed) into a new leaf submodule
tui_prompt_relay/idle_offset_resolution.rs. Bodies are byte-identical; only the
per-fn `#[cfg(unix)]` is lifted to the parent's gated `mod`/`use` decls and
visibility is set to the minimum.

resolve_idle_tail_start_offset + clamp_idle_tail_start_offset_to_committed are
pub(super) and re-imported by the parent's prod call sites;
claude_idle_response_start_offset_after_timestamp is re-imported only under
`#[cfg(all(unix, test))]` (its prod callers are now child-internal — avoids an
unused-import warning, mirroring the rehydration precedent); the module-internal
normalize_transcript_fallback_offset stays private.

tui_prompt_relay.rs 4376 -> 4301 prod LoC; giant baseline + change-surfaces.md
freeze synced to 4301; new child module is 100 prod lines (below the cap).
