+++ ARCHITECTURE.md
@@ -534,6 +534,7 @@ src/
 │   │   │   └── store.rs
 │   │   ├── tui_prompt_relay/
 │   │   │   ├── anchor_completion.rs
+│   │   │   ├── idle_offset_resolution.rs
 │   │   │   ├── idle_transcript_scan.rs
 │   │   │   ├── injected_prompt_policy.rs
 │   │   │   ├── launch_script.rs
